bugs fixing and changes 7th sep

+++ app/Http/Controllers/API/Contests/ContestController.php
@@ -58,17 +58,17 @@ class ContestController extends Controller
         DB::beginTransaction();
         try
         {
-            $user_package = UserPackageSubscription::where('user_id',Auth::id())->where('module','Contest')->orderBy('created_at','desc')->first();
-            if(empty($user_package))
-            {
-                return response()->json(prepareResult(true, ['No Package Subscribed'], getLangByLabelGroups('messages','message_no_package_subscribed_error')), config('http_response.internal_server_error'));
-            }
-            elseif($user_package->number_of_contest == $user_package->used_number_of_contest)
-            {
-                return response()->json(prepareResult(true, ['Package Use Exhausted'], getLangByLabelGroups('messages','message_job_ads_exhausted_error')), config('http_response.internal_server_error'));
-            }
-            else
-            {
+            // $user_package = UserPackageSubscription::where('user_id',Auth::id())->where('module','Contest')->orderBy('created_at','desc')->first();
+            // if(empty($user_package))
+            // {
+            //     return response()->json(prepareResult(true, ['No Package Subscribed'], getLangByLabelGroups('messages','message_no_package_subscribed_error')), config('http_response.internal_server_error'));
+            // }
+            // elseif($user_package->number_of_contest == $user_package->used_number_of_contest)
+            // {
+            //     return response()->json(prepareResult(true, ['Package Use Exhausted'], getLangByLabelGroups('messages','message_job_ads_exhausted_error')), config('http_response.internal_server_error'));
+            // }
+            // else
+            // {
                 $getLastContest = Contest::select('id')->orderBy('created_at','DESC')->first();
                 if($getLastContest) {
                     $contestNumber = $getLastContest->auto_id;
@@ -147,12 +147,12 @@ class ContestController extends Controller
                 $contest->meta_keywords                         = $request->meta_keywords;
                 $contest->meta_description                      = $request->meta_description;
                 $contest->is_deleted                    		= false;
-                // $contest->status                    			= $request->status;
+                $contest->status                    			= 'pending';
                 $contest->tags                         = json_encode($request->tags);
                 $contest->status                                = 'verified';
                 $contest->save();
 
-                $user_package->update(['used_number_of_contest'=>($user_package->used_number_of_contest + 1),'used_number_of_event'=>($user_package->used_number_of_event + 1)]);
+                // $user_package->update(['used_number_of_contest'=>($user_package->used_number_of_contest + 1),'used_number_of_event'=>($user_package->used_number_of_event + 1)]);
 
                 if(!empty($request->cancellation_ranges))
                 {
@@ -230,7 +230,7 @@ class ContestController extends Controller
 
                 // Notification End
                 
-            }
+            // }
             DB::commit();
             $contest = Contest::with('user:id,first_name,last_name,gender,dob,email,contact_number,profile_pic_path','categoryMaster','subCategory','addressDetail','cancellationRanges','user.serviceProviderDetail:id,user_id,company_logo_path')->find($contest->id);
             return response()->json(prepareResult(false, $contest, getLangByLabelGroups('messages','messages_contest_created')), config('http_response.created'));
@@ -273,7 +273,7 @@ class ContestController extends Controller
         {
             $is_abuse_reported = false;
         }
-        $contest = Contest::with('user:id,first_name,last_name,gender,dob,email,contact_number,profile_pic_path','categoryMaster','subCategory','cancellationRanges','user.serviceProviderDetail:id,user_id,company_logo_path','contestWinners')->withCount('contestApplications')->find($contest->id);
+        $contest = Contest::with('user:id,first_name,last_name,gender,dob,email,contact_number,profile_pic_path,show_email,show_contact_number','categoryMaster','subCategory','cancellationRanges','user.serviceProviderDetail:id,user_id,company_logo_path','contestWinners')->withCount('contestApplications')->find($contest->id);
         $contest['is_applied'] = $applied;
         $contest['auth_application'] = $authApplication;
 
@@ -402,14 +402,18 @@ class ContestController extends Controller
             	}
             }
 
-            ContestTag::where('contest_id',$contest->id)->delete();
-            foreach ($request->tags as $key => $tag) {
-                $allTypeTag = new ContestTag;
-                $allTypeTag->contest_id                 = $contest->id;
-                $allTypeTag->user_id                  = Auth::id();
-                $allTypeTag->title                    = $tag;
-                $allTypeTag->type                     = $request->type;
-                $allTypeTag->save();
+            if(!empty($request->tags) && is_array($request->tags))
+            {
+
+                ContestTag::where('contest_id',$contest->id)->delete();
+                foreach ($request->tags as $key => $tag) {
+                    $allTypeTag = new ContestTag;
+                    $allTypeTag->contest_id                 = $contest->id;
+                    $allTypeTag->user_id                  = Auth::id();
+                    $allTypeTag->title                    = $tag;
+                    $allTypeTag->type                     = $request->type;
+                    $allTypeTag->save();
+                }
             }
